+++ bigtable/integration_test.go
@@ -22,6 +22,7 @@ import (
 	"fmt"
 	"math"
 	"math/rand"
+	"os"
 	"os/exec"
 	"reflect"
 	"sort"
@@ -1113,7 +1114,6 @@ func TestIntegration_Admin(t *testing.T) {
 	}
 	if iAdminClient != nil {
 		defer iAdminClient.Close()
-
 		iInfo, err := iAdminClient.InstanceInfo(ctx, adminClient.instance)
 		if err != nil {
 			t.Errorf("InstanceInfo: %v", err)
@@ -1196,6 +1196,14 @@ func TestIntegration_Admin(t *testing.T) {
 		t.Errorf("Column family mismatch, got %v, want %v", tblInfo.Families, wantFams)
 	}
 
+	encInfo, err := adminClient.EncryptionInfo(ctx, tblConf.TableID)
+	if err != nil {
+		t.Errorf("Encryption Info does not expect err: %v", err)
+	}
+	if !testutil.Equal(len(encInfo), 0) {
+		t.Errorf("Encryption Info mismatch, got %v, want %v", len(encInfo), 0)
+	}
+
 	// Populate mytable and drop row ranges
 	if err = adminClient.CreateColumnFamily(ctx, "mytable", "cf"); err != nil {
 		t.Fatalf("Creating column family: %v", err)
@@ -1253,6 +1261,15 @@ func TestIntegration_Admin(t *testing.T) {
 	}))
 	if gotRowCount != 0 {
 		t.Errorf("Invalid row count after truncating table: got %v, want %v", gotRowCount, 0)
+	}
+
+	// Validate Encyrption Info not configured
+	encryptionInfo, err := adminClient.EncryptionInfo(ctx, "mytable")
+	if err != nil {
+		t.Fatalf("EncryptionInfo: %v", err)
+	}
+	if got, want := len(encryptionInfo), 0; !cmp.Equal(got, want) {
+		t.Fatalf("Number of Clusters with Encryption Info: %v, want: %v", got, want)
 	}
 }
 
@@ -1397,6 +1414,7 @@ func TestIntegration_AdminCreateInstance(t *testing.T) {
 	if err := iAdminClient.CreateInstance(ctx, conf); err != nil {
 		t.Fatalf("CreateInstance: %v", err)
 	}
+
 	defer iAdminClient.DeleteInstance(ctx, instanceToCreate)
 
 	iInfo, err := iAdminClient.InstanceInfo(ctx, instanceToCreate)
@@ -1449,6 +1467,175 @@ func TestIntegration_AdminCreateInstance(t *testing.T) {
 
 	if cInfo.ServeNodes != 5 {
 		t.Fatalf("NumNodes: %v, want: %v", cInfo.ServeNodes, 5)
+	}
+
+	if cInfo.KMSKeyName != "" {
+		t.Fatalf("KMSKeyName: %v, want: %v", cInfo.KMSKeyName, "")
+	}
+}
+
+func TestIntegration_AdminEncryptionInfo(t *testing.T) {
+	if instanceToCreate == "" {
+		t.Skip("instanceToCreate not set, skipping instance creation testing")
+	}
+
+	testEnv, err := NewIntegrationEnv()
+	if err != nil {
+		t.Fatalf("IntegrationEnv: %v", err)
+	}
+	defer testEnv.Close()
+
+	if !testEnv.Config().UseProd {
+		t.Skip("emulator doesn't support instance creation")
+	}
+
+	timeout := 5 * time.Minute
+	ctx, cancel := context.WithTimeout(context.Background(), timeout)
+	defer cancel()
+
+	iAdminClient, err := testEnv.NewInstanceAdminClient()
+	if err != nil {
+		t.Fatalf("NewInstanceAdminClient: %v", err)
+	}
+	defer iAdminClient.Close()
+
+	adminClient, err := testEnv.NewAdminClient()
+	if err != nil {
+		t.Fatalf("NewAdminClient: %v", err)
+	}
+	defer adminClient.Close()
+
+	table := testEnv.Config().Table
+	clusterID := testEnv.Config().Cluster
+
+	keyRingName := os.Getenv("GCLOUD_TESTS_GOLANG_KEYRING")
+	if keyRingName == "" {
+		t.Fatal("GCLOUD_TESTS_GOLANG_KEYRING must be set. See CONTRIBUTING.md for details")
+	}
+	kmsKeyName := keyRingName + "/cryptoKeys/key1"
+
+	conf := &InstanceWithClustersConfig{
+		InstanceID:  instanceToCreate,
+		DisplayName: "test instance",
+		Clusters: []ClusterConfig{
+			{
+				ClusterID:  clusterID,
+				KMSKeyName: kmsKeyName,
+				Zone:       instanceToCreateZone,
+				NumNodes:   1,
+			},
+		},
+	}
+	if err := iAdminClient.CreateInstanceWithClusters(ctx, conf); err != nil {
+		t.Fatalf("CreateInstance: %v", err)
+	}
+	defer iAdminClient.DeleteInstance(ctx, instanceToCreate)
+
+	// Delete the table at the end of the test. Schedule ahead of time
+	// in case the client fails
+	defer deleteTable(ctx, t, adminClient, table)
+	if err := adminClient.CreateTable(ctx, table); err != nil {
+		t.Fatalf("Creating table: %v", err)
+	}
+
+	encryptionKeyVersion := kmsKeyName + "/cryptoKeyVersions/1"
+
+	// The encryption info can take 30-300s (currently about 120-190s) to
+	// become ready.
+	for i := 0; i < 30; i++ {
+		encryptionInfo, err := adminClient.EncryptionInfo(ctx, table)
+		if err != nil {
+			t.Fatalf("EncryptionInfo: %v", err)
+		}
+
+		kmsKeyVersion := encryptionInfo[clusterID][0].KMSKeyVersion
+		if kmsKeyVersion != "" {
+			break
+		}
+
+		time.Sleep(time.Second * 10)
+	}
+
+	// Validate Encryption Info under getTable
+	table2, err := adminClient.getTable(ctx, table, btapb.Table_ENCRYPTION_VIEW)
+	if err != nil {
+		t.Fatalf("Getting Table: %v", err)
+	}
+	if got, want := len(table2.ClusterStates), 1; !cmp.Equal(got, want) {
+		t.Fatalf("Table Cluster States %v, want: %v", got, want)
+	}
+	clusterState := table2.ClusterStates[clusterID]
+	if got, want := len(clusterState.EncryptionInfo), 1; !cmp.Equal(got, want) {
+		t.Fatalf("Table Encryption Info Length: %v, want: %v", got, want)
+	}
+	tableEncInfo := clusterState.EncryptionInfo[0]
+	if got, want := int(tableEncInfo.EncryptionStatus.Code), 0; !cmp.Equal(got, want) {
+		t.Fatalf("EncryptionStatus: %v, want: %v", got, want)
+	}
+	// NOTE: this EncryptionType is btapb.EncryptionInfo_EncryptionType
+	if got, want := tableEncInfo.EncryptionType, btapb.EncryptionInfo_CUSTOMER_MANAGED_ENCRYPTION; !cmp.Equal(got, want) {
+		t.Fatalf("EncryptionType: %v, want: %v", got, want)
+	}
+	if got, want := tableEncInfo.KmsKeyVersion, encryptionKeyVersion; !cmp.Equal(got, want) {
+		t.Fatalf("KMS Key Version: %v, want: %v", got, want)
+	}
+
+	// Validate Encyrption Info retrieved via EncryptionInfo
+	encryptionInfo, err := adminClient.EncryptionInfo(ctx, table)
+	if err != nil {
+		t.Fatalf("EncryptionInfo: %v", err)
+	}
+	if got, want := len(encryptionInfo), 1; !cmp.Equal(got, want) {
+		t.Fatalf("Number of Clusters with Encryption Info: %v, want: %v", got, want)
+	}
+	encryptionInfos := encryptionInfo[clusterID]
+	if got, want := len(encryptionInfos), 1; !cmp.Equal(got, want) {
+		t.Fatalf("Encryption Info Length: %v, want: %v", got, want)
+	}
+	if len(encryptionInfos) != 1 {
+		t.Fatalf("Expected Single EncryptionInfo")
+	}
+	v := encryptionInfos[0]
+	if got, want := int(v.Status.Code), 0; !cmp.Equal(got, want) {
+		t.Fatalf("EncryptionStatus: %v, want: %v", got, want)
+	}
+	// NOTE: this EncryptionType is EncryptionType
+	if got, want := v.Type, CustomerManagedEncryption; !cmp.Equal(got, want) {
+		t.Fatalf("EncryptionType: %v, want: %v", got, want)
+	}
+	if got, want := v.KMSKeyVersion, encryptionKeyVersion; !cmp.Equal(got, want) {
+		t.Fatalf("KMS Key Version: %v, want: %v", got, want)
+	}
+
+	// Validate CMEK on Cluster Info
+	cInfo, err := iAdminClient.GetCluster(ctx, instanceToCreate, clusterID)
+	if err != nil {
+		t.Fatalf("GetCluster: %v", err)
+	}
+
+	if got, want := cInfo.KMSKeyName, kmsKeyName; !cmp.Equal(got, want) {
+		t.Fatalf("KMSKeyName: %v, want: %v", got, want)
+	}
+
+	// Create a backup with CMEK enabled, verify backup encryption info
+	backupName := "backupCMEK"
+	defer adminClient.DeleteBackup(ctx, clusterID, backupName)
+	if err = adminClient.CreateBackup(ctx, table, clusterID, backupName, time.Now().Add(8*time.Hour)); err != nil {
+		t.Fatalf("Creating backup: %v", err)
+	}
+	backup, err := adminClient.BackupInfo(ctx, clusterID, backupName)
+	if err != nil {
+		t.Fatalf("BackupInfo: %v", backup)
+	}
+
+	if got, want := backup.EncryptionInfo.Type, CustomerManagedEncryption; !cmp.Equal(got, want) {
+		t.Fatalf("Backup Encryption EncryptionType: %v, want: %v", got, want)
+	}
+	if got, want := backup.EncryptionInfo.KMSKeyVersion, encryptionKeyVersion; !cmp.Equal(got, want) {
+		t.Fatalf("Backup Encryption KMSKeyVersion: %v, want: %v", got, want)
+	}
+	if got, want := int(backup.EncryptionInfo.Status.Code), 2; !cmp.Equal(got, want) {
+		t.Fatalf("Backup EncryptionStatus: %v, want: %v", got, want)
 	}
 }
 
@@ -2254,12 +2441,43 @@ func TestIntegration_AdminBackup(t *testing.T) {
 	}
 	defer adminClient.Close()
 
-	table := testEnv.Config().Table
-	cluster := testEnv.Config().Cluster
-
+	tblConf := TableConf{
+		TableID: testEnv.Config().Table,
+		Families: map[string]GCPolicy{
+			"fam1": MaxVersionsPolicy(1),
+			"fam2": MaxVersionsPolicy(2),
+		},
+	}
+	if err := adminClient.CreateTableFromConf(ctx, &tblConf); err != nil {
+		t.Fatalf("Creating table from TableConf: %v", err)
+	}
 	// Delete the table at the end of the test. Schedule ahead of time
 	// in case the client fails
-	defer deleteTable(ctx, t, adminClient, table)
+	defer deleteTable(ctx, t, adminClient, tblConf.TableID)
+
+	sourceInstance := testEnv.Config().Instance
+	sourceCluster := testEnv.Config().Cluster
+
+	iAdminClient, err := testEnv.NewInstanceAdminClient()
+	if err != nil {
+		t.Fatalf("NewInstanceAdminClient: %v", err)
+	}
+	defer iAdminClient.Close()
+	diffInstance := testEnv.Config().Instance + "-diff"
+	diffCluster := sourceCluster + "-diff"
+	conf := &InstanceConf{
+		InstanceId:   diffInstance,
+		ClusterId:    diffCluster,
+		DisplayName:  "different test sourceInstance",
+		Zone:         instanceToCreateZone2,
+		InstanceType: DEVELOPMENT,
+		Labels:       map[string]string{"test-label-key": "test-label-value"},
+	}
+	defer iAdminClient.DeleteInstance(ctx, diffInstance)
+	// Create different instance to restore table.
+	if err := iAdminClient.CreateInstance(ctx, conf); err != nil {
+		t.Errorf("CreateInstance: %v", err)
+	}
 
 	list := func(cluster string) ([]*BackupInfo, error) {
 		infos := []*BackupInfo(nil)
@@ -2278,12 +2496,8 @@ func TestIntegration_AdminBackup(t *testing.T) {
 		return infos, err
 	}
 
-	if err := adminClient.CreateTable(ctx, table); err != nil {
-		t.Fatalf("Creating table: %v", err)
-	}
-
 	// Precondition: no backups
-	backups, err := list(cluster)
+	backups, err := list(sourceCluster)
 	if err != nil {
 		t.Fatalf("Initial backup list: %v", err)
 	}
@@ -2293,14 +2507,14 @@ func TestIntegration_AdminBackup(t *testing.T) {
 
 	// Create backup
 	backupName := "mybackup"
-	defer adminClient.DeleteBackup(ctx, cluster, "mybackup")
+	defer adminClient.DeleteBackup(ctx, sourceCluster, "mybackup")
 
-	if err = adminClient.CreateBackup(ctx, table, cluster, backupName, time.Now().Add(8*time.Hour)); err != nil {
+	if err = adminClient.CreateBackup(ctx, tblConf.TableID, sourceCluster, backupName, time.Now().Add(8*time.Hour)); err != nil {
 		t.Fatalf("Creating backup: %v", err)
 	}
 
 	// List backup
-	backups, err = list(cluster)
+	backups, err = list(sourceCluster)
 	if err != nil {
 		t.Fatalf("Listing backups: %v", err)
 	}
@@ -2308,62 +2522,93 @@ func TestIntegration_AdminBackup(t *testing.T) {
 		t.Fatalf("Listing backup count: %d, want: %d", got, want)
 	}
 	if got, want := backups[0].Name, backupName; got != want {
-		t.Fatalf("Backup name: %s, want: %s", got, want)
+		t.Errorf("Backup name: %s, want: %s", got, want)
 	}
-	if got, want := backups[0].SourceTable, table; got != want {
-		t.Fatalf("Backup SourceTable: %s, want: %s", got, want)
+	if got, want := backups[0].SourceTable, tblConf.TableID; got != want {
+		t.Errorf("Backup SourceTable: %s, want: %s", got, want)
 	}
 	if got, want := backups[0].ExpireTime, backups[0].StartTime.Add(8*time.Hour); math.Abs(got.Sub(want).Minutes()) > 1 {
-		t.Fatalf("Backup ExpireTime: %s, want: %s", got, want)
+		t.Errorf("Backup ExpireTime: %s, want: %s", got, want)
 	}
 
 	// Get backup
-	backup, err := adminClient.BackupInfo(ctx, cluster, backupName)
+	backup, err := adminClient.BackupInfo(ctx, sourceCluster, backupName)
 	if err != nil {
 		t.Fatalf("BackupInfo: %v", backup)
 	}
-	if got, want := *backup, *backups[0]; got != want {
-		t.Fatalf("BackupInfo: %v, want: %v", got, want)
+	if got, want := *backup, *backups[0]; cmp.Equal(got, &want) {
+		t.Errorf("BackupInfo: %v, want: %v", got, want)
 	}
 
 	// Update backup
 	newExpireTime := time.Now().Add(10 * time.Hour)
-	err = adminClient.UpdateBackup(ctx, cluster, backupName, newExpireTime)
+	err = adminClient.UpdateBackup(ctx, sourceCluster, backupName, newExpireTime)
 	if err != nil {
 		t.Fatalf("UpdateBackup failed: %v", err)
 	}
 
 	// Check that updated backup has the correct expire time
-	updatedBackup, err := adminClient.BackupInfo(ctx, cluster, backupName)
+	updatedBackup, err := adminClient.BackupInfo(ctx, sourceCluster, backupName)
 	if err != nil {
 		t.Fatalf("BackupInfo: %v", err)
 	}
 	backup.ExpireTime = newExpireTime
 	// Server clock and local clock may not be perfectly sync'ed.
 	if got, want := *updatedBackup, *backup; got.ExpireTime.Sub(want.ExpireTime) > time.Minute {
-		t.Fatalf("BackupInfo: %v, want: %v", got, want)
+		t.Errorf("BackupInfo: %v, want: %v", got, want)
 	}
 
 	// Restore backup
-	restoredTable := table + "-restored"
+	restoredTable := tblConf.TableID + "-restored"
 	defer deleteTable(ctx, t, adminClient, restoredTable)
-	if err = adminClient.RestoreTable(ctx, restoredTable, cluster, backupName); err != nil {
+	if err = adminClient.RestoreTable(ctx, restoredTable, sourceCluster, backupName); err != nil {
 		t.Fatalf("RestoreTable: %v", err)
 	}
 	if _, err := adminClient.TableInfo(ctx, restoredTable); err != nil {
 		t.Fatalf("Restored TableInfo: %v", err)
 	}
+	// Restore backup to different instance
+	restoreTableName := tblConf.TableID + "-diff-restored"
+	diffConf := IntegrationTestConfig{
+		Project:  testEnv.Config().Project,
+		Instance: diffInstance,
+		Cluster:  diffCluster,
+		Table:    restoreTableName,
+	}
+	env := &ProdEnv{
+		config: diffConf,
+	}
+	dAdminClient, err := env.NewAdminClient()
+	if err != nil {
+		t.Errorf("NewAdminClient: %v", err)
+	}
+	defer dAdminClient.Close()
+
+	defer deleteTable(ctx, t, dAdminClient, restoreTableName)
+	if err = dAdminClient.RestoreTableFrom(ctx, sourceInstance, restoreTableName, sourceCluster, "mybackup"); err != nil {
+		t.Fatalf("RestoreTableFrom: %v", err)
+	}
+	tblInfo, err := dAdminClient.TableInfo(ctx, restoreTableName)
+	if err != nil {
+		t.Fatalf("Restored to different sourceInstance failed, TableInfo: %v", err)
+	}
+	families := tblInfo.Families
+	sort.Strings(tblInfo.Families)
+	wantFams := []string{"fam1", "fam2"}
+	if !testutil.Equal(families, wantFams) {
+		t.Errorf("Column family mismatch, got %v, want %v", tblInfo.Families, wantFams)
+	}
 
 	// Delete backup
-	if err = adminClient.DeleteBackup(ctx, cluster, backupName); err != nil {
+	if err = adminClient.DeleteBackup(ctx, sourceCluster, backupName); err != nil {
 		t.Fatalf("DeleteBackup: %v", err)
 	}
-	backups, err = list(cluster)
+	backups, err = list(sourceCluster)
 	if err != nil {
 		t.Fatalf("List after Delete: %v", err)
 	}
 	if got, want := len(backups), 0; got != want {
-		t.Fatalf("List after delete len: %d, want: %d", got, want)
+		t.Errorf("List after delete len: %d, want: %d", got, want)
 	}
 }